feat: Authentication, Users & permissions
- Send email
- Forgot password
- Reset password
- Update user details
- Admin Users CRUD

+++ controllers/bootcamps.js
@@ -35,7 +35,24 @@ exports.getBootcamp = asyncHandler(async (req, res, next) => {
 // @route   POST /api/v1/bootcamps
 // @access  Private
 exports.createBootcamp = asyncHandler(async (req, res, next) => {
+  // Add user to req.body
+  req.body.user = req.user.id;
+
+  // Check for published bootcamp
+  const publishedBootcamp = await Bootcamp.findOne({ user: req.user.id });
+
+  // If the user is not an admin, they can only add one bootcamp
+  if (publishedBootcamp && req.user.role !== "admin") {
+    return next(
+      new ErrorResponse(
+        `The user with ID ${req.user.id} has already published a bootcamp`,
+        400
+      )
+    );
+  }
+
   const bootcamp = await Bootcamp.create(req.body);
+
   res.status(201).json({
     success: true,
     data: bootcamp,
@@ -46,10 +63,7 @@ exports.createBootcamp = asyncHandler(async (req, res, next) => {
 // @route   PUT /api/v1/bootcamps/:id
 // @access  Private
 exports.updateBootcamp = asyncHandler(async (req, res, next) => {
-  const bootcamp = await Bootcamp.findByIdAndUpdate(req.params.id, req.body, {
-    new: true,
-    runValidator: true,
-  });
+  let bootcamp = await Bootcamp.findById(req.params.id);
 
   if (!bootcamp) {
     return next(
@@ -57,10 +71,27 @@ exports.updateBootcamp = asyncHandler(async (req, res, next) => {
     );
   }
 
-  res.status(200).json({
-    success: true,
-    data: bootcamp,
+  // Make sure user is bootcamp owner
+  if (bootcamp.user.toString() !== req.user.id && req.user.role !== "admin") {
+    return next(
+      new ErrorResponse(
+        `User ${req.user.id} is not authorized to update this bootcamp`,
+        401
+      )
+    );
+  }
+
+  // update slug while updating name
+  if (Object.keys(req.body).includes("name")) {
+    req.body.slug = slugify(req.body.name, { lower: true });
+  }
+
+  bootcamp = await Bootcamp.findByIdAndUpdate(req.params.id, req.body, {
+    new: true,
+    runValidators: true,
   });
+
+  res.status(200).json({ success: true, data: bootcamp });
 });
 
 // @desc    Delete bootcamp
@@ -75,13 +106,19 @@ exports.deleteBootcamp = asyncHandler(async (req, res, next) => {
     );
   }
 
-  // Cascade delete courses when a bootcamp is deleted
-  await bootcamp.deleteOne();
+  // Make sure user is bootcamp owner
+  if (bootcamp.user.toString() !== req.user.id && req.user.role !== "admin") {
+    return next(
+      new ErrorResponse(
+        `User ${req.user.id} is not authorized to delete this bootcamp`,
+        401
+      )
+    );
+  }
 
-  res.status(200).json({
-    success: true,
-    data: {},
-  });
+  bootcamp.deleteOne();
+
+  res.status(200).json({ success: true, data: {} });
 });
 
 // @desc    Get bootcamps within a radius
@@ -122,15 +159,25 @@ exports.bootcampPhotoUpload = asyncHandler(async (req, res, next) => {
     );
   }
 
+  // Make sure user is bootcamp owner
+  if (bootcamp.user.toString() !== req.user.id && req.user.role !== "admin") {
+    return next(
+      new ErrorResponse(
+        `User ${req.user.id} is not authorized to update this bootcamp`,
+        401
+      )
+    );
+  }
+
   if (!req.files) {
-    return next(new ErrorResponse(`Please upload a file`, 404));
+    return next(new ErrorResponse(`Please upload a file`, 400));
   }
 
   const file = req.files.file;
 
   // Make sure the image is a photo
   if (!file.mimetype.startsWith("image")) {
-    return next(new ErrorResponse(`Please upload an image file`, 404));
+    return next(new ErrorResponse(`Please upload an image file`, 400));
   }
 
   // Check filesize
@@ -138,7 +185,7 @@ exports.bootcampPhotoUpload = asyncHandler(async (req, res, next) => {
     return next(
       new ErrorResponse(
         `Please upload an image less than ${process.env.MAX_FILE_UPLOAD}`,
-        404
+        400
       )
     );
   }
@@ -149,7 +196,7 @@ exports.bootcampPhotoUpload = asyncHandler(async (req, res, next) => {
   file.mv(`${process.env.FILE_UPLOAD_PATH}/${file.name}`, async (err) => {
     if (err) {
       console.error(err);
-      return next(new ErrorResponse(`Problem with file upload `, 500));
+      return next(new ErrorResponse(`Problem with file upload`, 500));
     }
 
     await Bootcamp.findByIdAndUpdate(req.params.id, { photo: file.name });
@@ -159,6 +206,4 @@ exports.bootcampPhotoUpload = asyncHandler(async (req, res, next) => {
       data: file.name,
     });
   });
-
-  console.log(file.name);
 });
